Se separo el codigo a una interfaz para que sea mas escalable
