+++ src/app/pages/listapacientes/listapacientes-page.component.html
@@ -95,7 +95,7 @@
           <div class="subvista">
             <app-info-paciente
               *ngIf="subvista === 'info'"
-              [paciente]="pacienteSeleccionado"
+              [paciente]="pacienteInfo"
               [propietario]="propietarioSeleccionado">
             </app-info-paciente>
